pyproject.toml
[project]
name = "news-clustering-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "nltk>=3.9.2",
    "pandas>=2.3.3",
    "plotly>=6.5.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "streamlit>=1.52.2",
]
